update(core): remove old task class
