+++ packages/core/src/index.ts
@@ -2,6 +2,6 @@ export * from './event-emitter';
 export * from './logger';
 export * from './manifest';
 export * from './project';
-export * from './task';
+export * from './tasks';
 export * from './task-manager';
 export * from './workspace';
